blogger platform(2)

+++ src/domain/blogs/use-cases/delete-post-with-checking-use-case.ts
@@ -27,7 +27,8 @@ export class DeletePostWithCheckingUseCase
 
     const blog = await this.blogsRepository.findBlogById(command.blogId);
     if (!blog) return HttpStatus.NOT_FOUND;
-    // if (blog && blog.userId !== command.userId) return HttpStatus.FORBIDDEN;
+    if (blog && blog.user && blog.user.id !== command.userId)
+      return HttpStatus.FORBIDDEN;
 
     const result = await this.postsRepository.deletePost(command.postId);
